requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Pillow==10.1.0
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
stripe==7.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
